# docs/v1/swagger.yaml
basePath: /v1
definitions:
...
  request.AnalysisUrlDocReq:
    properties:
      knowledgeId:
        type: string
      urlList:
        items:
          type: string
        type: array
    required:
    - knowledgeId
    - urlList
    type: object
...
  request.BatchCreateDocSegmentReq:
    properties:
      docId:
        description: 文档id
        type: string
      fileUploadId:
        description: fileUploadId
        type: string
    required:
    - docId
    - fileUploadId
    type: object
  request.BatchDocMetaDataReq:
    properties:
      createMeta:
        description: 文档没设置过对应key则创建元数据
        type: boolean
      knowledgeId:
        type: string
      metaDataList:
        description: 文档元数据
        items:
          $ref: '#/definitions/request.DocMetaData'
        type: array
    type: object
...
  request.ChildChunk:
    properties:
      chunkNo:
        description: 子分段序列号
        type: integer
      content:
        description: 子分段内容
        type: string
    type: object
...
  request.CreateDocChildSegmentReq:
    properties:
      content:
        description: 分段内容
        items:
          type: string
        type: array
      docId:
        description: 文档id
        type: string
      parentId:
        description: 父分段id
        type: string
    required:
    - content
    - docId
    - parentId
    type: object
  request.CreateDocSegmentReq:
    properties:
      content:
        description: 分段内容
        type: string
      docId:
        description: 文档id
        type: string
      labels:
        description: 关键词列表
        items:
          type: string
        type: array
    required:
    - content
    - docId
    - labels
    type: object
...
  request.DeleteDocChildSegmentReq:
    properties:
      ChildChunkNoList:
        description: 子分段序列号列表
        items:
          type: integer
        type: array
      docId:
        description: 文档id
        type: string
      parentChunkNo:
        description: 父分段序列号
        type: integer
      parentId:
        description: 父分段id
        type: string
    required:
    - ChildChunkNoList
    - docId
    - parentId
    type: object
  request.DeleteDocReq:
    properties:
      docIdList:
        items:
          type: string
        type: array
      knowledgeId:
        type: string
    required:
    - docIdList
    - knowledgeId
    type: object
  request.DeleteDocSegmentReq:
    properties:
      contentId:
        type: string
      docId:
        description: 文档id
        type: string
    required:
    - contentId
    - docId
    type: object
...
  request.DocImportReq:
    properties:
      docAnalyzer:
        description: 文档解析类型 text / ocr  / model
        items:
          type: string
        type: array
      docImportType:
        description: 文档导入类型，0：文件上传，1：url上传，2.批量url上传
        type: integer
      docInfoList:
        description: 上传文档列表
        items:
          $ref: '#/definitions/request.DocInfo'
        type: array
      docMetaData:
        description: 元数据
        items:
          $ref: '#/definitions/request.DocMetaData'
        type: array
      docPreprocess:
        description: 文本预处理规则 replaceSymbols / deleteLinks
        items:
          type: string
        type: array
      docSegment:
        allOf:
        - $ref: '#/definitions/request.DocSegment'
        description: 文档分段配置
      knowledgeId:
        description: 知识库id
        type: string
      parserModelId:
        description: 模型解析或ocr模型id
        type: string
    required:
    - docAnalyzer
    - docInfoList
    - docSegment
    - knowledgeId
    type: object
  request.DocInfo:
    properties:
      docId:
        description: 文档id
        type: string
      docName:
        description: 文档名称
        type: string
      docSize:
        description: 文档类型
        type: integer
      docType:
        description: 文档类型
        type: string
      docUrl:
        description: 文档url
        type: string
    type: object
  request.DocMetaData:
    properties:
      metaId:
        description: 元数据id
        type: string
      metaKey:
        description: key
        type: string
      metaRule:
        description: 正则表达式
        type: string
      metaValue:
        description: 确定值
        type: string
      metaValueType:
        description: string，number，time
        type: string
      option:
        description: option:add(新增)、update(更新)、delete(删除),update 和delete 的时候metaId
          不能为空
        type: string
    type: object
  request.DocMetaDataReq:
    properties:
      docId:
        type: string
      knowledgeId:
        type: string
      metaDataList:
        description: 文档元数据
        items:
          $ref: '#/definitions/request.DocMetaData'
        type: array
    type: object
  request.DocSegment:
    properties:
      maxSplitter:
        description: 可分隔最大值（只有自定义分段必填）
        type: integer
      overlap:
        description: 可重叠值（只有自定义分段必填）
        type: number
      segmentMethod:
        description: 分段方法 0：通用分段；1：父子分段
        type: string
      segmentType:
        description: 分段方式，只有通用分段必填 0：自动分段；1：自定义分段
        type: string
      splitter:
        description: 分隔符（只有自定义分段必填）
        items:
          type: string
        type: array
      subMaxSplitter:
        description: 可分隔最大值（只有父子分段必填）
        type: integer
      subSplitter:
        description: 分隔符（只有父子分段必填）
        items:
          type: string
        type: array
    required:
    - segmentMethod
    type: object
  request.DocSegmentLabelsReq:
    properties:
      contentId:
        type: string
      docId:
        type: string
      labels:
        items:
          type: string
        type: array
    required:
    - contentId
    - docId
    - labels
    type: object
...
  request.UpdateDocChildSegmentReq:
    properties:
      childChunk:
        allOf:
        - $ref: '#/definitions/request.ChildChunk'
        description: 子分段序列号列表
      docId:
        description: 文档id
        type: string
      parentChunkNo:
        description: 父分段序列号
        type: integer
      parentId:
        description: 父分段id
        type: string
    required:
    - childChunk
    - docId
    - parentId
    type: object
  request.UpdateDocSegmentReq:
    properties:
      content:
        type: string
      contentId:
        type: string
      docId:
        type: string
    required:
    - content
    - contentId
    - docId
    type: object
  request.UpdateDocSegmentStatusReq:
    properties:
      all:
        description: all 代表全部启用，此时将忽略contentId
        type: boolean
      contentId:
        type: string
      contentStatus:
        type: string
      docId:
        type: string
    required:
    - contentStatus
    - docId
    type: object
...
  response.AnalysisDocUrlResp:
    properties:
      urlList:
        items:
          $ref: '#/definitions/response.Url'
        type: array
    type: object
...
  response.ChildSegmentInfo:
    properties:
      childId:
        description: 子分段id
        type: string
      childNum:
        description: 子分段序号
        type: integer
      content:
        description: 内容
        type: string
      parentId:
        description: 父分段id
        type: string
    type: object
...
  response.DocChildSegmentResp:
    properties:
      contentList:
        description: 内容
        items:
          $ref: '#/definitions/response.ChildSegmentInfo'
        type: array
    type: object
...
  response.DocMetaData:
    properties:
      metaId:
        description: 元数据id
        type: string
      metaKey:
        description: key
        type: string
      metaRule:
        description: 正则表达式
        type: string
      metaValue:
        description: 确定值
        type: string
      metaValueType:
        description: number，time, string
        type: string
    type: object
...
  response.DocSegmentResp:
    properties:
      contentList:
        description: 内容
        items:
          $ref: '#/definitions/response.SegmentContent'
        type: array
      fileName:
        description: 名称
        type: string
      maxSegmentSize:
        description: 设置最大长度
        type: integer
      metaDataList:
        description: 文档元数据
        items:
          $ref: '#/definitions/response.DocMetaData'
        type: array
      pageTotal:
        description: 总页数
        type: integer
      segmentImportStatus:
        description: 分段导入状态描述
        type: string
      segmentMethod:
        description: 分段方式 父子分段/通用分段
        type: string
      segmentTotalNum:
        description: 分段数量
        type: integer
      segmentType:
        description: 分段方式 0自动分段 1自定义分段
        type: string
      splitter:
        description: 分隔符（只有自定义分段必填）
        type: string
      uploadTime:
        description: 上传时间
        type: string
    type: object
...
  response.ListDocResp:
    properties:
      docId:
        type: string
      docName:
        description: 文档名称
        type: string
      docType:
        description: 文档类型
        type: string
      errorMsg:
        description: 解析错误信息，预留
        type: string
      fileSize:
        description: 文件大小，预留
        type: string
      knowledgeId:
        description: 知识库id
        type: string
      segmentMethod:
        description: 分段模式0:通用分段，1：父子分段
        type: string
      status:
        description: 处理状态
        type: integer
      uploadTime:
        description: 上传时间
        type: string
    type: object
...
  response.PageResult:
    properties:
      list: {}
      pageNo:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  response.SegmentContent:
    properties:
      available:
        type: boolean
      childNum:
        description: 子分段数量
        type: integer
      content:
        type: string
      contentId:
        type: string
      contentNum:
        type: integer
      isParent:
        description: 父子分段/通用分段 true是父分段，false是通用分段
        type: boolean
      labels:
        items:
          type: string
        type: array
    type: object
...
  response.Url:
    properties:
      fileName:
        type: string
      fileSize:
        type: integer
      url:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    ## HTTP Header
    | Header        | 说明      |
    | ------------- | --------- |
    | Authorization | JWT token |
    | X-Language    | 语言Code  |
    | X-Org-Id      | 组织ID    |
    | X-Client-Id   | 客户端标识|

    ## HTTP Status
    | HTTP Status             | 说明                   |
    | ----------------------- | ---------------------- |
    | 200, StatusOK           | 请求返回成功           |
    | 400, StatusBadRequest   | 请求返回失败，用于业务 |
    | 401, StatusUnauthorized | JWT认证失败            |
    | 403, StatusForbidden    | 没有权限               |

    ## 权限-菜单对应表
    | 一级权限        | 二级权限  | 三级权限 | 一级菜单 | 二级菜单 | 三级菜单 |
    |-------------|-------|------|------|------|------|
    | guest       |       |      | 【访客】 |      |      |
    | common      |       |      | 【通用】 |      |      |
    | permission  |       |      | 权限管理 |      |      |
    | permission  | user  |      | 权限管理 | 用户管理 |      |
    | permission  | org   |      | 权限管理 | 组织管理 |      |
    | permission  | role  |      | 权限管理 | 角色管理 |      |

    ## `/v1/user/permission`返回用例
    ```json
    {
      "code": 0,
      "data": {
        "orgPermission": {
          "org": {"id": "test-org-id", "name": "test-org-name"},
          "permissions": [
            {"perm": "permission"},
            {"perm": "permission.user"},
            {"perm": "permission.org"},
            {"perm": "permission.role"}
          ]
        }
      },
      "msg": "操作成功"
    }
    ```
...
paths:
...
  /knowledge/doc:
    delete:
      consumes:
      - application/json
      description: 删除文档
      parameters:
      - description: 删除文档请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DeleteDocReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 删除文档
      tags:
      - knowledge
  /knowledge/doc/import:
    post:
      consumes:
      - application/json
      description: 上传文档
      parameters:
      - description: 文档上传请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DocImportReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 上传文档
      tags:
      - knowledge
  /knowledge/doc/import/tip:
    get:
      consumes:
      - application/json
      description: 获取知识库异步上传任务提示：有正在执行的异步上传任务/最近一次上传任务的失败信息
      parameters:
      - in: query
        name: knowledgeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 获取知识库异步上传任务提示
      tags:
      - knowledge
  /knowledge/doc/list:
    get:
      consumes:
      - application/json
      description: 获取知识库文档列表，不展示状态为无效（-1）的文档数据
      parameters:
      - in: query
        name: docName
        type: string
      - in: query
        name: knowledgeId
        required: true
        type: string
      - in: query
        name: pageNo
        type: integer
      - in: query
        name: pageSize
        required: true
        type: integer
      - description: 当前状态  -1-全部， 0-待处理， 1- 处理完成， 2-正在审核中，3-正在解析中，4-审核未通过，5-解析失败
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.PageResult'
            - properties:
                list:
                  items:
                    $ref: '#/definitions/response.ListDocResp'
                  type: array
              type: object
      security:
      - JWT: []
      summary: 获取文档列表
      tags:
      - knowledge
  /knowledge/doc/meta:
    post:
      consumes:
      - application/json
      description: 更新文档元数据
      parameters:
      - description: 文档更新元数据请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DocMetaDataReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 更新文档元数据
      tags:
      - knowledge
  /knowledge/doc/meta/batch:
    post:
      consumes:
      - application/json
      description: 批量更新文档元数据
      parameters:
      - description: 批量文档更新元数据请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.BatchDocMetaDataReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 批量更新文档元数据
      tags:
      - knowledge
  /knowledge/doc/segment/batch/create:
    post:
      consumes:
      - application/json
      description: 批量新增文档切片
      parameters:
      - description: 批量新增文档切片请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.BatchCreateDocSegmentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 批量新增文档切片
      tags:
      - knowledge
  /knowledge/doc/segment/child/create:
    post:
      consumes:
      - application/json
      description: 新增文档子分片
      parameters:
      - description: 新增文档子切片请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CreateDocChildSegmentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 新增文档子分片
      tags:
      - knowledge
  /knowledge/doc/segment/child/delete:
    delete:
      consumes:
      - application/json
      description: 删除文档子切片
      parameters:
      - description: 删除文档切片请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DeleteDocChildSegmentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 删除文档子切片
      tags:
      - knowledge
  /knowledge/doc/segment/child/list:
    get:
      consumes:
      - application/json
      description: 获取子分段列表
      parameters:
      - in: query
        name: contentId
        required: true
        type: string
      - in: query
        name: docId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.DocChildSegmentResp'
              type: object
      security:
      - JWT: []
      summary: 获取子分段列表
      tags:
      - knowledge
  /knowledge/doc/segment/child/update:
    post:
      consumes:
      - application/json
      description: 更新文档子切片
      parameters:
      - description: 更新文档子切片请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UpdateDocChildSegmentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 更新文档子切片
      tags:
      - knowledge
  /knowledge/doc/segment/create:
    post:
      consumes:
      - application/json
      description: 新增文档切片
      parameters:
      - description: 新增文档切片请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CreateDocSegmentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 新增文档切片
      tags:
      - knowledge
  /knowledge/doc/segment/delete:
    delete:
      consumes:
      - application/json
      description: 删除文档切片
      parameters:
      - description: 删除文档切片请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DeleteDocSegmentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 删除文档切片
      tags:
      - knowledge
  /knowledge/doc/segment/labels:
    post:
      consumes:
      - application/json
      description: 更新文档切片标签
      parameters:
      - description: 更新文档切片标签请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DocSegmentLabelsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 更新文档切片标签
      tags:
      - knowledge
  /knowledge/doc/segment/list:
    get:
      consumes:
      - application/json
      description: 获取文档切分结果
      parameters:
      - in: query
        name: docId
        required: true
        type: string
      - in: query
        name: pageNo
        type: integer
      - in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.DocSegmentResp'
              type: object
      security:
      - JWT: []
      summary: 获取文档切分结果
      tags:
      - knowledge
  /knowledge/doc/segment/status/update:
    post:
      consumes:
      - application/json
      description: 更新文档切片启用状态
      parameters:
      - description: 更新文档切片启用状态请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UpdateDocSegmentStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 更新文档切片启用状态
      tags:
      - knowledge
  /knowledge/doc/segment/update:
    post:
      consumes:
      - application/json
      description: 更新文档切片
      parameters:
      - description: 更新文档切片请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UpdateDocSegmentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 更新文档切片
      tags:
      - knowledge
  /knowledge/doc/url/analysis:
    post:
      consumes:
      - application/json
      description: 解析url
      parameters:
      - description: 解析url请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AnalysisUrlDocReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.AnalysisDocUrlResp'
              type: object
      security:
      - JWT: []
      summary: 解析url
      tags:
      - knowledge
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"